// frontend/src/api/generated/models/materiaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { HorarioMateriaDto } from './horarioMateriaDto';
import type { MateriaDtoEstado } from './materiaDtoEstado';
import type { ProfessorDto } from './professorDto';
import type { TurmaResumoDto } from './turmaResumoDto';

export interface MateriaDto {
  id: string;
  nome: string;
  turmaId: string;
  professorId: string;
  cargaHorariaReferencia: number;
  horarios: HorarioMateriaDto[];
  estado: MateriaDtoEstado;
  /** @nullable */
  encerradaEm: string | null;
  /** Decimal serializado como string — média mínima pra aprovação, padrão 7 */
  notaMinimaAprovacao: string;
  turma: TurmaResumoDto;
  professor: ProfessorDto;
}